Fix VpcId dict key in ec2_eip and add tests (#181)
* Fix VpcId dict key in ec2_eip and add tests

We're not converting camel_to_snake in this module, however ensure_present
has been using snake_cased keys for vpc_id. This prevents attaching an EIP
to an instance. This appears to have been overlooked in the boto3 migration
in ansible/ansible/pull/61575.

Also adds tests to cover the `if is_instance` codepath in ensure_present.