ec2_instance: Fix spurious error message when we lose a race (#7)
It is possible for all instances to stop matching the filters between
the initial check for existing instances and the first call to
find_instances() in change_instance_state().

If this happened, find_instances() would previously be called a second
time with an empty list of instance IDs and no filters, which should
not happen and immediately ends module execution with the error "No
filters provided when they were required".