ec2_instance: clarify module state vs EC2 state
In some cases, we need to branch based on desired
EC2 instance state. In others, on desired module state.

Some CI checks were failing due to not checking the
right value. Renamed and/or assigned local variables
in a few critical functions here to be clear on which
state we are referring to.