ec2_vol tests set requires to == rather than >= (so we can test the minimum we claim) (#453)
ec2_vol tests set botocore requires to == rather than >=

SUMMARY
We currently use >= which generally means that the latest version of botocore will be pulled in.  Given that we specify it's supposed to work with >= we should test the oldest possible version to ensure that we support at least that version.
ISSUE TYPE

Tests Pull Request

COMPONENT NAME
ec2_vol
ADDITIONAL INFORMATION
Depends-on: #460

Reviewed-by: None <None>