Support changing of launch type (#840)
Support changing of launch type

SUMMARY
When changing the launch_type parameter for an ecs_taskdefition there was no change reported by the module. This adds a check to see if launch_type in the ecs_taskdefinition has changed. If there is a change detected the module reports back there is not a matching task definition and creates a new one.
ISSUE TYPE

Bugfix Pull Request

COMPONENT NAME
ecs_taskdefinition

Reviewed-by: Alina Buzachis <None>
Reviewed-by: Michael Mayer <mjmayer@gmail.com>
Reviewed-by: Markus Bergholz <git@osuv.de>