Fix on_denied and on_missing bugs (#618)
Fix on_denied and on_missing bugs

SUMMARY
This pull request:

Changes the default value of on_denied to be error, so that it agrees with what is stated in the documentation.
Changes the default value of on_missing to be error, and updates the documentation to explain this.

Fixes #617.
ISSUE TYPE

Bugfix Pull Request

COMPONENT NAME
aws_ssm lookup

Reviewed-by: Alina Buzachis <None>
Reviewed-by: Shane Frasier <maverick@maverickdolphin.com>
Reviewed-by: Markus Bergholz <git@osuv.de>