Module naming guidelines (#881)
Module naming guidelines

SUMMARY
Originally discussed in #610
Here's an initial attempt at some module naming (and module scoping) guidelines.
Assuming we adopt this, I recommend is that:

we can do the renames as part of 5.0.0, but this isn't a must and wouldn't be considered a blocker to 5.0.0 unless otherwise agreed.
any modules moving from community.aws to amazon.aws would be renamed if appropriate as part of the move.
we do not specify a deprecation date at this time, but leave a comment in the redirect that any deprecation cycle should not start before 2024-09-01 (to avoid potential migration fatigue issues),

ISSUE TYPE

Docs Pull Request

COMPONENT NAME
docs/docsite/rst/dev_guidelines.rst
ADDITIONAL INFORMATION
fixes: #610

Reviewed-by: Joseph Torcasso <None>
Reviewed-by: Jill R <None>
Reviewed-by: Markus Bergholz <git@osuv.de>