s3_object: Add parameter `acl_disabled` to handle uploading files to buckets with ACL disabled. (#921)
s3_object: Add parameter `acl_disabled` to handle uploading files to buckets with ACL disabled.

SUMMARY

Fixes #863
[Update: 07/07/2022]
https://boto3.amazonaws.com/v1/documentation/api/latest/reference/services/s3.html#S3.Client.get_bucket_ownership_controls can be used to determine if ACL is disabled or not.
Modified code to convert acl_disabled from a user input parameter to a variable used for testing if ACL is enabled/disabled.
Add parameter  acl_disabled to handle uploading files to buckets with ACL disabled.
If set to true, all the permission related operations are skipped in module code.
ISSUE TYPE


Bugfix Pull Request

COMPONENT NAME

s3_object
ADDITIONAL INFORMATION


AWS added the option to create S3 bucket with ACL Disabled in Nov 2021) and made it the default/suggested setting when creating S3 bucket through AWS Portal .
Currently the ACL "permission" parameter defaults to ["private"] and there is no way to tell the s3_object module to omit the ACL setting while uploading files to a bucket which has set the ACL as disabled.
Tried looking for a way to get the existing bucket info to determine if 'ACL' is enabled/disabled, but was not able to find what I was looking for in API documentation.

Reviewed-by: Alina Buzachis <None>
Reviewed-by: Jill R <None>
Reviewed-by: Mark Chappell <None>
Reviewed-by: Mandar Kulkarni <mandar242@gmail.com>