Handle absence of classic link support for shared VPCs (#984)
Handle absence of classic link support for shared VPCs

Signed-off-by: GomathiselviS gomathiselvi@gmail.com
SUMMARY


Fixes #663
One of the Classic Link Limitations is that
You cannot link an EC2-Classic instance to a VPC that's in a different Region or a different AWS account..
When calling the DescribeVpcClassicLink operation on a shared VPC, it errors out saying "The vpc ID 'vpc-xxxx' does not exist".
This PR handles this error.
ISSUE TYPE


Bugfix Pull Request

COMPONENT NAME

ADDITIONAL INFORMATION

Reviewed-by: Mark Chappell <None>