inventory/aws_ec2: deprecate include_extra_api_calls (#1097)
inventory/aws_ec2: deprecate include_extra_api_calls

The include_extra_api_calls option of the inventory plugin is broken
and actually crashes the plugin with a list index out of range error.
It assumes describe_instance_status always returns a list with one entry
which is not the  case if the instance is stopped.
In addition, it reads the event from for a Events key that doesn't exist
in the answer structure.
https://boto3.amazonaws.com/v1/documentation/api/latest/reference/services/ec2.html#EC2.Client.describe_instance_status
The feature is not covered by any tests in our CI and obviously, nobody uses it.

Reviewed-by: Alina Buzachis <None>
Reviewed-by: Mark Chappell <None>