Deprecate module_utils.urls (#1146)
Deprecate module_utils.urls

SUMMARY
Originally introduced in ansible/ansible#42758, as far as I can tell the module_utils/urls.py has never actually been used by a module.  It's primary purpose was to add support for Sigv4, but boto3/botocore support this out of the box now and we shouldn't be reimplementing it.
We have no tests for this code, nothing's using it, and it's in the "supported" repo.  Let's prune out this code, if someone external's using it hopefully someone will raise an issue and we can reconsider the deprecation.
ISSUE TYPE

Feature Pull Request

COMPONENT NAME
plugins/module_utils/urls.py
ADDITIONAL INFORMATION

Reviewed-by: Alina Buzachis <None>