Update version_added information for modules which migrate to amazon.aws in 5.0.0 (#1155)
Update version_added information for modules which migrate to amazon.aws in 5.0.0

SUMMARY
Felix has pointed out in IRC/Matrix that for modules which migrated to amazon.aws it can be a little confusing when you try to look back in logs to find exactly what needs to be installed for a feature (installing amazon.aws.4.3.0 doesn't guarantee that community.aws 4.x is installed)
Update the main 'version_added' to match the version that the modules were moved to amazon.aws, and add the version_added_collection for existing new features.
ISSUE TYPE

Docs Pull Request

COMPONENT NAME
various modules
ADDITIONAL INFORMATION
Scope is currently limited to the 5.0.0 migrations, to simplify backports.  If accepted I'll cleanup the 4.0.0 and 2.0.0 modules too.

Reviewed-by: Felix Fontein <felix@fontein.de>