Add documentation for the use of strings with compose (#1283)
Add documentation for the use of strings with compose

SUMMARY
Documentation for the use of strings with compose is documented in https://docs.ansible.com/ansible/latest/plugins/inventory.html#using-inventory-plugins. However, in the module, it is not well documented. This documentation and example provides documentation in the module docs.
Related to #570
ISSUE TYPE

Docs Pull Request

COMPONENT NAME
aws_ec2 inventory

Reviewed-by: Gonéri Le Bouder <goneri@lebouder.net>
Reviewed-by: Mark Chappell <None>