Add support to AnsibleAWSModule.client to override the default parameters (#1303)
Add support to AnsibleAWSModule.client to override the default parameters

Depends-On: ansible-collections/community.aws#1635
SUMMARY
When refactoring the inventory plugins it was helpful to be able to override region, however looking at the S3 code, it's also helpful to be able to override other things too.  Pull this in as a small change while I work on the rest of it.
ISSUE TYPE

Feature Pull Request

COMPONENT NAME
plugins/module_utils/modules.py
ADDITIONAL INFORMATION

Reviewed-by: Alina Buzachis <None>