s3_bucket - be more forgiving when missing 'get' permissions for a parameter we're not trying to set (#1406)
s3_bucket - be more forgiving when missing 'get' permissions for a parameter we're not trying to set

SUMMARY
S3 permissions are painfully complex.  Be as forgiving as possible when missing a 'get' permission on a parameter we're not trying to set.
ISSUE TYPE

Feature Pull Request

COMPONENT NAME
s3_bucket
ADDITIONAL INFORMATION
Related to #1352

Reviewed-by: Alina Buzachis