[PR #1359/31570048 backport] [stable-5] ec2_metadata_facts - fix AttributeError when running on older managed hosts (#1420)
[PR #1359/31570048 backport] [stable-5] ec2_metadata_facts - fix AttributeError when running on older managed hosts

This is a MANUAL backport of PR #1359 as merged into main (3157004).
SUMMARY
Unlike most of our modules, the ec2_metadata_facts is always expected to run on managed hosts, and doesn't depend on the AWS SDKs. With Python 2.7 support still in place for managed hosts on all supported Ansible versions it makes sense, in the specific case of ec2_metadata_facts, to continue to support Python 2.7 for now.
ISSUE TYPE
Bugfix Pull Request
COMPONENT NAME
ec2_metadata_facts
ADDITIONAL INFORMATION
fixes: #1358