Backup vault tag bugfix (#1610)
Backup vault tag bugfix

SUMMARY
Fixes an error raised in backup_vault when updating tags on an existing vault by calling the correct boto3 client methods for tagging and untagging a resource. This is not applicable when creating a vault as tags are included in the create_backup_vault method params, but when updating a vault we have to call the tag methods individually.
ISSUE TYPE

Bugfix Pull Request

COMPONENT NAME
backup_vault
ADDITIONAL INFORMATION
See the included new integration test task for a command that previously failed and works with the updated code.

Reviewed-by: Alina Buzachis
Reviewed-by: Helen Bailey <hebailey@redhat.com>
Reviewed-by: Mike Graves <mgraves@redhat.com>