Add ability to remove detached internet gateway (#1786)
Add ability to remove detached internet gateway

SUMMARY

Introduces the ability to remove an internet gateway that is not attached to a VPC. To remove an internet gateway either the ID of the internet gateway or the ID of the attached VPC can be supplied. It is also possible to supply both IDs. If both IDs are supplied then a failure will be generated if the attached VPC ID does not match the user supplied VPC.

Fixes #1669
ISSUE TYPE


Feature Pull Request

COMPONENT NAME

ec2_vpc_igw
ADDITIONAL INFORMATION



I wasn't able to add a test for removing an internet gateway that is detached as the module currently doesn't have a way to create a detached internet gateway. I plan on opening a follow-up PR for that use case and will add an appropriate removal test in that PR.

Reviewed-by: Mark Chappell
Reviewed-by: Brant Evans <bevans@redhat.com>
Reviewed-by: Brian A. Teller