Avoid unsafe conditionals in integration tests (#1891)
Avoid unsafe conditionals in integration tests

SUMMARY
Update a batch of assertions in integration tests to avoid templating
See: ansible/ansible-documentation#866 and ansible/ansible-documentation#866 for more information
This PR covers about half of the impacted test tasks for this repo but is already going to be a lot to get through CI so additional PRs will follow.
Affected tasks identified with:
https://gist.github.com/sivel/7b012ae2ebf4ab4cb6225379a72cc502
ISSUE TYPE

Bugfix Pull Request

COMPONENT NAME
tests
ADDITIONAL INFORMATION
There are unsupported tests in this change.  The cloudtrail tests seem to be broken independent of this change, iam_group tests pass for me locally

Reviewed-by: Bikouo Aubin
Reviewed-by: Mark Chappell
Reviewed-by: Alina Buzachis
Reviewed-by: Mike Graves <mgraves@redhat.com>