Fix missing FQCNs in integration test (#1902) (#1906)
[stable-7] Fix missing FQCNs in integration test (#1902)

SUMMARY
This PR fixes the integration tests to use FQCN throughout. Every change has been generated by running ansible-lint --fix with the write_list limited to fqcn. The one exception is the manual removal of the tests/integration/targets/aws_region_info/main.yml which ansible-lint was failing on. This seems to be unused? It attempts to include tasks/tests.yml which does not exist.
The ansible-lint config used to generate this changeset:
---
exclude_paths:
  - plugins/ write_list:
  - fqcn

I had to exclude the plugins/ directory because ansible-lint was crashing, I suspect because of docstrings. It did not provide enough information in debug or traceback to identify which file was causing the problem.
ISSUE TYPE

Bugfix Pull Request

COMPONENT NAME
ADDITIONAL INFORMATION

Reviewed-by: Mike Graves <mgraves@redhat.com>
Reviewed-by: Alina Buzachis