module_utils/elbv2 - fix issue with authenticate-oidc listener rule (#1956)
module_utils/elbv2 - fix issue with authenticate-oidc listener rule

SUMMARY

fixes #1877
The module now detect rule which changing priority, they are no more considered as new rules but we are using the set_rule_priorities API instead to update the priority.
For authenticated-oidc rule, we set always set the UseExistingSecret to False for new rule to create and when the rule need to be modified and the user has provided a ClientSecret.

ISSUE TYPE


Bugfix Pull Request

Reviewed-by: Helen Bailey <hebailey@redhat.com>
Reviewed-by: Bikouo Aubin
Reviewed-by: Alina Buzachis