update ec2_eip integration tests (#2198)
SUMMARY

The validation of the tests consists in checking all the existing EIP from the account, this is unstable as it may generate unexpected if another process is running and create EIP in the same time.
We are fixing by adding tag with the resource_prefix to all EIP created and restrict EIP comparison to the ones having this tag.

ISSUE TYPE


Bugfix Pull Request

Reviewed-by: Rahmanim Benny <brahmani@redhat.com>
Reviewed-by: Alina Buzachis
Reviewed-by: Helen Bailey <hebailey@redhat.com>