cloudwatchlogs_log_group_metric_filter: add support for unit and dimensions (#2286)
SUMMARY

This PR adds support for the options unit and dimensions in the cloudwatchlogs_log_group_metric_filter module.
This enables configuring unit and dimensions in Cloudwatch Logs Metricfilters using ansible, which was previously not possible.
The addition is pretty straigthforward since both unit and dimensions are part of the metric_transformation parameter.
dimensions and default_value are mutually exclusive, however:

The AWS API does not fail when both are present, instead the dimensions are simply ignored.
Since they are not top-level parameters, the ansible module option mutually_exclusive was not possible. Instead a custom check was added to the module that throws an error when both parameters are present.

An integration test has been added for this case, as well as for configuring metric_filters with units and/or dimensions.
The function metricTransformationHandler has been rewritten slightly due to the addition of the two extra optional parameters, to make it a bit more readable.
Happy to get your feedback!

ISSUE TYPE

Feature Pull Request

COMPONENT NAME

cloudwatchlogs_log_group_metric_filter
ADDITIONAL INFORMATION

Reviewed-by: GomathiselviS <gomathiselvi@gmail.com>
Reviewed-by: Jasper Misset
Reviewed-by: Alina Buzachis
(cherry picked from commit 118905b8004526bfc8289ba41e0ac3d053d347c6)