Switch inventory_aws_ec2 over to using setup_ec2_facts for pulling AMI info (#2448)
PR #2397 added some integration test cleanup that is needed in the stable-8 branch to fix the failing integration tests in PR #2435. However the entire PR can't be backported to stable-8 because it also includes some updates that are only relevant to the stable-9 branch. This backport PR includes only the commit from #2397 with the necessary integration test updates.

Reviewed-by: Mark Chappell