Add generate_mark() function and corresponding unit test. (#2235)
SUMMARY

I created a function inside of the Connection class to generate a random string to replace existing repetitive code. Since this functionality was only currently used in this particular class and still relies on the MARK_LENGTH variable, I did not pull it out as a separate/common utility.
This function generates a random string of characters based on a globally-defined MARK_LENGTH variable, which seem to be used to bookend SSM CLI input.
This existed in the previous iteration of this code, and I did not change it because I was not able to ascertain why 26 was chosen as the length in the first place, and I did not want to make a prescriptive decision around that. There is an existing unit test that calls "5" as the MARK_LENGTH, so I assume that would have worked as well, but did not want to make breaking changes to the existing unit test. This may be something for a future TODO if the team wants to clarify/standardize a length for this string.
I wrote a unit test for this and confirmed that it passes, though I mainly did that for my own benefit and practice, and don't believe that this function really required any unit testing given what it's doing.
Fixes #ACA-2096
ISSUE TYPE

Feature Pull Request

COMPONENT NAME

Function to generate a random string for AWS SSM CLI  delimiting.

Reviewed-by: Mark Chappell
Reviewed-by: Mandar Kulkarni <mandar242@gmail.com>
Reviewed-by: GomathiselviS <gomathiselvi@gmail.com>
Reviewed-by: Bianca Henderson <beeankha@gmail.com>
Reviewed-by: Bikouo Aubin
Reviewed-by: Mike Graves <mgraves@redhat.com>

This commit was initially merged in https://github.com/ansible-collections/community.aws
See: https://github.com/ansible-collections/community.aws/commit/bde08f92fb187fe1a789847bd670fca46ad50b3f