Add new workflow to label prs with needs_triage (#2783)
SUMMARY

This pr adds a new workflow for labeling new and reopened prs that are not marked as draft. The needs_triage label will be removed if the pr is marked as draft during development and re-added once the pr is marked as ready for review.

After consulting with the team, we decided to label prs in a new workflow to allow for the prs and issues to have different labels in the future.

ACA-2362
ISSUE TYPE

    Feature Pull Request

TESTING

Verified the needs_triage label is added on a new pr in hdefazio#3